pyproject.toml
[project]
name = "fastapi-ordering-ddd"
version = "0.1.0"
description = "Simple ordering system with FastAPI, Clean Architecture, PostgreSQL and RabbitMQ"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.0.0",
    "aio-pika>=9.0.0",
    "alembic>=1.14.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: testes de API (HTTP)",
    "integration_db: testes de repositório com PostgreSQL",
    "unit: testes de domínio e aplicação",
]

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"alembic/env.py" = ["E402"]
